+++ app/artifact-bundle/page.tsx
@@ -1,0 +1,5 @@
+import ArtifactBundleZipPanel from "../../components/ArtifactBundleZipPanel";
+
+export default function ArtifactBundlePage() {
+  return <ArtifactBundleZipPanel />;
+}